+++ tests/bootstrap.php
@@ -9,4 +9,5 @@
  */
 
 require_once __DIR__.'/../includes/ReportType.php';
+require_once __DIR__.'/../includes/ReportComparison.php';
 require_once __DIR__.'/../includes/reports/ItemPercentilesReport.php';